fix(MediaWikiConfigReader): ->has() should not fall back to PHP settings
`getConfigByVariableName` is still falling back to return the main PHP
settings if the requested config name cannot be found. That is (for now)
still the behavior we want for `->get()`. However for `->has()` it
should just respond with `false` if the requested config name is not
available in any CommunityConfiguration provider.

`getConfigByVariableName` is logging a (debug) message if there is no
CommunityConfig for the requested config name. While that warning is the
behavior we want if the config is actually requested by `->get()`,
checking whether a config exists with `->has()` should cause no such
message. Especially, since we want to replace that message with an
exception eventually.

This should have been part of I92627bad64 (commit 91787af) for T387452.

Bug: T387452
Change-Id: I2f57fa44eacfc899dcadcdc29af45ae7a759f66e