refactor: throw exception if setting does not exist in community config
That mechanism was briefly still used in 1.45.0-wmf.2 and 1.45.0-wmf.3
by the CampaignEvents extension for the `CampaignEventsEnableEventTypes`
variable. This has been fixed in
I935d1fe4d37a14c2bef495056e08977de5e9ef94.

Now that there is no usage anymore of that fallback-mechanism in
1.45.0-wmf.4 and onward, we can move forward to replacing it with
throwing an exception to prevent unintentional re-introduction of that.

If fallback to PHP settings is needed, then MediaWikiConfigRouter is
intended for that.

Change-Id: If0e52465787387e20bb7790cd9f553cf468f37b7