Revert "MediaWikiConfigReader: Lower logging level"
This reverts commit bbec309e9d8984606a75f74009064b4aadc2dc1a.

Reason for revert: We now implemented T387452, and as far as I can see, there should be no more instances where it is falling back to reading from the PHP settings. Thus raising the severity back to info to see if we missed any. A follow-up change will replace this with an exception.

Change-Id: Id1fbd4911b4f048f95b82db9c7d465674c6aef07