refactor: make IValidator implementations version-aware
While the JsonSchemaValidator now support version-aware validation, this
is not yet actually used anywhere. These usages will come with follow-up
changes, once also implementations of IValidator in other extensions and
the interface itself have been made version-aware.

Having version-aware validators is needed to be able to load and write
configuration data for older versions of a Schema, which in turn is
needed to handle migrations properly.

This also adds a test for the new functionality, and hardens the types
of the test case and drop a now pointless test.

Bug: T371028
Change-Id: I140d2505e813fcf6e80adc3a9e92ccbe44a3bdd5