ProviderFactory: Switch to ObjectFactory for provider creation
This fully respects all of ObjectFactory parameters. Specifically,
it properly implements `services` and `args`. Previously, the code
was using those attributes from an incorrect place (the provider
spec, rather than the provider _class_ spec).

Furthermore, this patch renames `$spec` to `$providerSpec`, making it
more clear what kind of specification it is, making this kind of mistake
more easier to spot.

Bug: T393233
Change-Id: I3ef1ab4acb4ebbc2466c4237b54463e303b152a5