tests: Remove unused function from EditorCapabilityFactoryTest
The function named "provideNewCapability" is not used in this test.
It's name indicates a data provider, but data provider cannot use
MainConfig to get the specs.

Use local variables to avoid repeated creation/get
of factory, context or title.
Remove assert, because EditorCapabilityFactory::newCapability has a type
declaration of the tested instance. The test runs the constructor of the
Capabilities to find failures, not assert return types.

Bug: T393079
Change-Id: I2054d36dfeb2063182432a4c47385ba4d59fa991