AbstractProvider: Ensure config is validated
Follow-up to I0ad19af1c15308fae73762256d99316202c543b1.

Michael noticed AbstractProvider no longer validates
the configuration. This was not an intentional change;
the goal of I0ad19af1c15308fae73762256d99316202c543b1
was to move normalization only, not the validation.

This is because AbuseFilter uses store format
which does not have an array on its top. In order
to be able to use AbstractProvider (with its own validator),
AbstractProvider should not force anything regarding the config
value.

Bug: T393240
Change-Id: I4bd5403ddb35b6625cd1c13c4e8a5d60d50bf33a