chore: update eslint-config-wikimedia to unbreak LibUp
With version 0.31.0, eslint-config-wikimedia now includes the typescript
linting-configuration that we formerly got from
@wmde/eslint-config-wikimedia-typescript. Also, these rules are now
automatically applied to all *.ts files.

This should now hopefully unbreak LibUp, which was so far blocked on
version conflicts between @wmde/eslint-config-wikimedia-typescript and
eslint-config-wikimedia.

The addition of cypress/.eslintrc.json was needed for that config to be
picked up. But why that file also needed "wikimedia/server" to actually
check any .ts files, that I do not understand. But now the cypress/
directory is still covered by eslint + typescript.

Change-Id: I839420b7f51983e5990e8c892cbb41efe32feaf9