Implement domain events interfaces
This locks in the method signature and name in place,
avoiding accidental mistakes. It also ensures other
developers implementing domain events implement
the right interface, as that is the current
recommended method of using them.

Last but not least, this actually makes it clear
from the listeners that they are using a deprecated
listener.

Bug: T401527
Change-Id: Ic5d693d89ab8395fbe037856208311c79511e0e4