fix: WikiPageStore's cache should be invalidated upon page deletion
Before this patch, WikiPageStore's cache was not invalidated
when a page was deleted. This is because we were looking only
for revision changes, and no revision change occurs in a page
deletion.

This patch fixes the problem by subscribing to PageDeletedEvent
as well. Originally, I wanted to subscribe PageRecordChangedEvent
(which is documented on MediaWiki.org) to catch even more events,
but that doesn't seem to work in practice (T403068).

Bug: T401322
Bug: T401529
Change-Id: I32487f446f40e10b3cead88ec63bb6890cea853b