refactor: Introduce ProviderServicesContainer
The container will be used to inject services into the provider
without having to change its signature across all children (which
may be in other extensions).

This is separate, to ensure the constructor signature changes
can be reviewed easily. See the following patch for actual
usage and reasoning.

The commit uses reflection to detect what signature is in
use. This adds a performance penalty, but this code will
be removed once the two affected extensions are fixed.
Relevant patches are listed as dependencies in Gerrit.

Bug: T405961
Change-Id: I7d532f8058f8cd8c086d5ba4e696a606916709b1