Move sorting values into DataProvider
Previously, sorting keys was included in the Writer. However, that
makes the Writer overly opinionated (it has to assume the top level
type is an object).

AbuseFilter uses a different top level type (an array), and keeping
the sorting in the Writer is incompatible with that.

Move this logic into the DataProvider. Since AbuseFilter inherits
AbstractProvider directly, this will fix its issue.

Bug: T406145
Change-Id: Ie7f6a3f5b1e19d89a46eeda9de61c12708552e8b