cypress: skip when CommunityConfigurationExample is not loaded
On CI, we would like to remove the recursive injection of dependencies.
CommunityConfiguration uses CommunityConfigurationExample which would no
more be added as a dependency for extensions solely depending on
ConfigurationConfiguration for testing (for example Cite or
GrowthExperiments).

I had PHPUnit test suite to conditionally skip with 2d0aaa5bc9e. This is
the follow up for Cypress.

The before hook queries the MediaWiki API to check whether
CommunityConfigurationExample is installed. When it is not, use Mocha
skip function.

I have tested it locally with and without CommunityConfigurationExample
being loaded.

Bug: T404939
Change-Id: I18390b904e0c4531d08702073d3b0774af6bd9d0